Changed from int to int64_t

+++ main.c
@@ -6,7 +6,7 @@ int main()
     str_strip(full_name);
     str_append(full_name, " Martin");
 
-    int start = 7;
+    int64_t start = 7;
     string_t *last_name = str_alloc_substr(full_name, &start, NULL, NULL);
 
     str_erase(full_name, "tin");
